Move handler definitions to a module
