--- backend/src/infra.rs
@@ -1,3 +1,0 @@
-pub mod user;
-
-pub use user::UserRepository;
