+++ backend/src/model.rs
@@ -2,4 +2,4 @@ pub mod key_pair;
 pub mod user;
 
 pub use key_pair::{KeyPair, KeyPairRepository};
-pub use user::User;
+pub use user::{User, UserRepository};
